Backport of DNS v2 - split up router into multiple responsibilities & break up router tests into multiple files. into release/1.18.x (#20780)
* backport of commit 008451aa1e05a0f224279d092f805fefaf1a3f45

* backport of commit abdf8c8205eddafcfdc2c7f94e4fef664f7da04a

* backport of commit a0176356f73f4724a44985471d29b85f3dc05b77

* DNS v2 - split up router into multiple responsibilities & break up router tests into multiple files. (#20688)

* Update agent/dns.go


* PR feedback

* split tests out into multiple files.

* Extract responsibilities from router into discoveryResultsFetcher, messageSerializer, responseGenerator.

* adding recordmaker tests

* add response generator test coverage.

* changing tests case name based on PR feedback

---------


---------
