+++ agent/dns/discovery_results_fetcher_test.go
